requirements.txt
torch==2.0.1
torchvision==0.15.2
gin-config==0.5.0
einops==0.6.1